refactor: split user list into components
